reanme and relocate countingGrid function
