+++ src/components/Pool.js
@@ -1,7 +1,7 @@
 import React from "react";
 import Piece from "./Piece";
 import DragShadow from "./DragShadow";
-import {countingGrid} from "./Board";
+import {getLetterCountPerSquare} from "../logic/getLetterCountPerSquare";
 
 export default function Pool({pieces, dragDestination, dispatchGameState}) {
   const poolPieces = pieces.filter((piece) => piece.poolIndex >= 0);
@@ -29,9 +29,11 @@ export default function Pool({pieces, dragDestination, dispatchGameState}) {
         <div className="pool-slot shadow" key={piece.id}>
           <DragShadow
             key={`shadow-piece-${piece.id}`}
-            grid={countingGrid(piece.letters.length, piece.letters[0].length, [
-              {...piece, dragGroupTop: 0, dragGroupLeft: 0},
-            ])}
+            grid={getLetterCountPerSquare(
+              piece.letters.length,
+              piece.letters[0].length,
+              [{...piece, dragGroupTop: 0, dragGroupLeft: 0}],
+            )}
           />
         </div>
       )),
